Split default/unauthenticated store data into separate file.
